+++ diagnosticism/__init__.py
@@ -9,9 +9,11 @@ __email__       =   'matthew@synesis.com.au'
 __license__     =   'BSD-3-Clause'
 __maintainer__  =   'Matt Wilson'
 __status__      =   'Beta'
-__version__     =   '0.2.2'
+__version__     =   '0.3.0'
 
 from .conrep import abort, report
+from .log import enable_logging, is_logging_enabled, log
 from .program_name import *
+from .severity import *
 from .trace import enable_tracing, is_tracing_enabled, trace